requirements.txt
acres==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
entrypoints==0.4
etelemetry==0.3.1
faiss-cpu==1.9.0.post1
Faker==33.1.0
fastapi==0.115.6
favicon==0.7.0
filelock==3.16.1
fitz==0.0.1.dev2
fonttools==4.55.3
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.7.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-ollama==0.2.1
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langserve==0.3.0
langsmith==0.1.147
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.16.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==1.26.4
ollama==0.4.4
openai==1.57.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
plotly==5.24.1
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.1
prov==2.0.1
puremagic==1.28
pyarrow==18.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.3
Pygments==2.18.0
pymdown-extensions==10.12
PyMuPDF==1.25.1
pyparsing==3.2.0
pypdf==5.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
st-annotated-text==4.0.1
st-theme==1.2.3
starlette==0.41.3
streamlit==1.40.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==6.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
validators==0.34.0
yarl==1.18.3
